apps/dashboard/requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic_core==2.20.1
